fix(engine): correct stale docs, cursor-mode log UB, FrameContext field order
- io/resource docs described the removed AssetId/findById path; assets
  resolve by name now (findByName)
- WindowManager doc called it a singleton; it is engine-owned
- drop misleading [[maybe_unused]] from the vec2/vec4 JSON helpers (used)
- CursorMode passed to a %d log specifier -> cast to int (-Wformat / UB)
- FrameContext field order now matches its own doc (deltaTime, simDeltaTime,
  fixedDeltaTime); engine.cpp's positional init reordered to match
- RenderView::environment doc said RenderSystem; it is copied from the Scene

+++ src/engine/platform/window/window_manager.cpp
@@ -229,7 +229,7 @@ void WindowManager::setCursorMode(CursorMode mode) {
             glfwmode = GLFW_CURSOR_CAPTURED;
             break;
         default:
-            LOG_ERROR("Invalid cursor mode: %d", mode);
+            LOG_ERROR("Invalid cursor mode: %d", static_cast<int>(mode));
             return;
     }
     glfwSetInputMode(windowContext, GLFW_CURSOR, glfwmode);
